Update SALEM_EA_B.mq4
The EA calculates and sets a single TP/SL to all the BUY/SELL orders (excluding pending orders) of the current pair so that their resultant profit will equal the "TP_DOLLARS" input set by the user. 
This EA eliminates the need to run an EA 24/7 on VPS that closes all the orders when their accumulative profit touches the input set by the user.